+++ src/database/Examverse/Table/users.sql
@@ -4,7 +4,7 @@
 
 CREATE TABLE IF NOT EXISTS public.users
 (
-    id integer NOT NULL DEFAULT nextval('users_id_seq'::regclass),
+    id bigint NOT NULL DEFAULT nextval('users_id_seq'::regclass),
     name character varying(100) COLLATE pg_catalog."default" NOT NULL,
     email character varying(100) COLLATE pg_catalog."default",
     mobile character varying(20) COLLATE pg_catalog."default" NOT NULL,
